Added shell.c that wraps all the file explorer code. Updated make file to include new source files.
