+++ sw/n64/testrom/shell.h
@@ -1,0 +1,1 @@
+void start_shell();
